feat: add server-side code with node to make server side requests
This was needed due to the fact that commercejs only accepts server side
requests in order to create a customer due to the scope. The serverjs file
contains the code that creates the server to accept incoming post requests
from the client side and then forwards that to the appropriate commerce api.